+++ app/bootstrap/factory.py
@@ -8,36 +8,41 @@ from fastapi import FastAPI
 from fastapi.staticfiles import StaticFiles
 from starlette.middleware.sessions import SessionMiddleware
 
+from app import config
+from app.db import init_auth_db
+from app.logging import configure_logging
+from app.middleware import auth_and_metrics_middleware
+from app.state import set_app
+
 
 @asynccontextmanager
 async def lifespan(_app: FastAPI):
     """Initialize application state for the FastAPI lifespan."""
-    from app import app as app_module
-
-    app_module.init_auth_db()
+    init_auth_db()
     yield
 
 
 def create_app() -> FastAPI:
     """Build and configure the Schwinn FastAPI application."""
-    from app import app as app_module
     from app.routes import account, admin, auth, grafana, health, workouts
 
     fastapi_app = FastAPI(title="Schwinn", version="1.0.0", lifespan=lifespan)
-    fastapi_app.secret_key = app_module.secret_key
-    fastapi_app.logger = app_module.logging.getLogger("schwinn")
+    fastapi_app.secret_key = config.secret_key
+    import logging
 
-    app_module.app = fastapi_app
-    app_module.configure_logging()
+    fastapi_app.logger = logging.getLogger("schwinn")
 
-    if app_module.STATIC_DIR.exists():
-        fastapi_app.mount("/static", StaticFiles(directory=str(app_module.STATIC_DIR)), name="static")
+    set_app(fastapi_app)
+    configure_logging()
 
-    fastapi_app.middleware("http")(app_module.auth_and_metrics_middleware)
+    if config.STATIC_DIR.exists():
+        fastapi_app.mount("/static", StaticFiles(directory=str(config.STATIC_DIR)), name="static")
+
+    fastapi_app.middleware("http")(auth_and_metrics_middleware)
     fastapi_app.add_middleware(
         SessionMiddleware,
-        secret_key=app_module.secret_key,
-        https_only=app_module.SESSION_COOKIE_SECURE,
+        secret_key=config.secret_key,
+        https_only=config.SESSION_COOKIE_SECURE,
         same_site="lax",
         session_cookie="session",
     )
