+++ app/db.py
@@ -1,0 +1,100 @@
+"""SQLite persistence helpers."""
+
+from __future__ import annotations
+
+import sqlite3
+from contextlib import contextmanager
+
+from app import config
+from app.state import get_logger
+
+
+@contextmanager
+def get_db_connection():
+    """Open a SQLite connection with row access."""
+    config.DATA_DIR.mkdir(parents=True, exist_ok=True)
+    connection = sqlite3.connect(config.AUTH_DB_FILE)
+    connection.row_factory = sqlite3.Row
+    try:
+        yield connection
+    finally:
+        connection.close()
+
+
+def init_auth_db() -> None:
+    """Create or migrate authentication tables and default settings."""
+    with get_db_connection() as connection:
+        connection.execute(
+            """
+            CREATE TABLE IF NOT EXISTS users (
+                id INTEGER PRIMARY KEY AUTOINCREMENT,
+                name TEXT NOT NULL DEFAULT '',
+                first_name TEXT NOT NULL DEFAULT '',
+                last_name TEXT NOT NULL DEFAULT '',
+                email TEXT NOT NULL UNIQUE,
+                email_verified INTEGER NOT NULL DEFAULT 0,
+                password_hash TEXT NOT NULL,
+                avatar_data BLOB,
+                avatar_mime TEXT NOT NULL DEFAULT '',
+                role TEXT NOT NULL DEFAULT 'user',
+                created_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
+            )
+            """
+        )
+        existing_columns = {
+            row["name"] for row in connection.execute("PRAGMA table_info(users)").fetchall()
+        }
+        if "name" not in existing_columns:
+            connection.execute("ALTER TABLE users ADD COLUMN name TEXT NOT NULL DEFAULT ''")
+        if "first_name" not in existing_columns:
+            connection.execute("ALTER TABLE users ADD COLUMN first_name TEXT NOT NULL DEFAULT ''")
+        if "last_name" not in existing_columns:
+            connection.execute("ALTER TABLE users ADD COLUMN last_name TEXT NOT NULL DEFAULT ''")
+        if "email_verified" not in existing_columns:
+            connection.execute("ALTER TABLE users ADD COLUMN email_verified INTEGER NOT NULL DEFAULT 0")
+        if "avatar_data" not in existing_columns:
+            connection.execute("ALTER TABLE users ADD COLUMN avatar_data BLOB")
+        if "avatar_mime" not in existing_columns:
+            connection.execute("ALTER TABLE users ADD COLUMN avatar_mime TEXT NOT NULL DEFAULT ''")
+        if "role" not in existing_columns:
+            connection.execute("ALTER TABLE users ADD COLUMN role TEXT NOT NULL DEFAULT 'user'")
+        connection.execute(
+            """
+            CREATE TABLE IF NOT EXISTS settings (
+                key TEXT PRIMARY KEY,
+                value TEXT NOT NULL
+            )
+            """
+        )
+        connection.execute(
+            "INSERT OR IGNORE INTO settings (key, value) VALUES (?, ?)",
+            ("registration_enabled", "false"),
+        )
+        connection.commit()
+
+
+def get_setting(key: str, default: str = "") -> str:
+    """Read an application setting from the database."""
+    with get_db_connection() as connection:
+        row = connection.execute("SELECT value FROM settings WHERE key = ?", (key,)).fetchone()
+    if row is None:
+        return default
+    return str(row["value"])
+
+
+def set_setting(key: str, value: str) -> bool:
+    """Create or update an application setting."""
+    try:
+        with get_db_connection() as connection:
+            connection.execute(
+                """
+                INSERT INTO settings (key, value) VALUES (?, ?)
+                ON CONFLICT(key) DO UPDATE SET value = excluded.value
+                """,
+                (key, value),
+            )
+            connection.commit()
+    except sqlite3.Error:
+        get_logger().warning("Settings update failed: key=%s", key)
+        return False
+    return True
