+++ app/routes/health.py
@@ -1,11 +1,20 @@
-"""Health and metrics route registrations."""
+"""Health and metrics routes."""
 
 from __future__ import annotations
 
-from fastapi import APIRouter
-
-from app import app as handlers
+from fastapi import APIRouter, Response
+from prometheus_client import CONTENT_TYPE_LATEST, generate_latest
 
 router = APIRouter()
-router.add_api_route("/healthz", handlers.healthz, methods=["GET"], name="healthz")
-router.add_api_route("/metrics", handlers.metrics, methods=["GET"], name="metrics")
+
+
+@router.get("/healthz", name="healthz")
+def healthz():
+    """Return application health status."""
+    return {"status": "ok"}
+
+
+@router.get("/metrics", name="metrics")
+def metrics():
+    """Return Prometheus metrics."""
+    return Response(generate_latest(), media_type=CONTENT_TYPE_LATEST)
