+++ app/services/__init__.py
@@ -1,0 +1,1 @@
+"""Service layer modules for Schwinn."""
