+++ app/workouts/__init__.py
@@ -1,0 +1,1 @@
+"""Workout parsing, analytics, and chart helpers."""
